+++ src/lib/command/CommandId.hpp
@@ -50,6 +50,7 @@ enum class CommandId : wxWindowID {
     FindPrevious,
     GotoLine,
     Format,
+    Reformat,
     Result,
     Browser,
     Minimap,